chg: Renamed the mapping module (which is no longer only a dictionaries file)
