+++ app/files/scripts/stix2/misp2stix2_mapping.py
